# Conflicts:
#	sen2agri-dashboard/dashboard.php

+++ sen2agri-processors/DEM-WB/test/demmaccs.py
@@ -65,6 +65,10 @@ parser.add_argument('--gip-dir', required=True, help="directory where gip are to
 parser.add_argument('--maccs-address', required=True, help="the ip address of the pc where MACCS is to be found")
 parser.add_argument('--maccs-launcher', required=True, help="the shell from the remote pc which launches the MACCS")
 parser.add_argument('--dem-launcher', required=True, help="executable for DEM")
+parser.add_argument('--skip-dem', required=False,
+                        help="skip DEM if a directory with previous work of DEM is given", default=None)
+parser.add_argument('--delete-temp', required=False,
+                        help="if set to True, it will delete all the temporary files and directories. Default: True", default="True")
 parser.add_argument('output', help="output location")
 
 args = parser.parse_args()
@@ -74,8 +78,10 @@ if not create_recursive_dirs(args.output):
     sys.exit(-1)
 
 general_log_path = args.output
-own_pid = os.getpid()
-working_dir = "{}/{}".format(args.working_dir,own_pid)
+
+working_dir = "{}/{}".format(args.working_dir[:len(args.working_dir) - 1] if args.working_dir.endswith("/") else args.working_dir, os.getpid())
+if args.skip_dem is not None:
+    working_dir = "{}".format(args.skip_dem)
 dem_working_dir = "{}_DEM_TMP".format(working_dir)
 dem_output_dir = "{}_DEM_OUT".format(working_dir)
 
@@ -83,6 +89,7 @@ log(general_log_path,"working_dir = {}".format(working_dir), general_log_filenam
 log(general_log_path,"dem_working_dir = {}".format(dem_working_dir), general_log_filename)
 log(general_log_path,"dem_output_dir = {}".format(dem_output_dir), general_log_filename)
 log(general_log_path, "Started at {}:".format(time.time()), general_log_filename)
+
 general_start = time.time()
 
 if not create_recursive_dirs(dem_output_dir):
@@ -94,9 +101,10 @@ if not create_recursive_dirs(working_dir):
     sys.exit(-1)
 
 start = time.time()
-if run_command([args.dem_launcher, "--srtm", args.srtm, "--swbd", args.swbd, "-p", args.processes_number_dem, "-w", dem_working_dir, args.input, dem_output_dir]) != 0:
-    log(general_log_path, "DEM failed", general_log_filename)
-    sys.exit(-1)
+if args.skip_dem is None:
+    if run_command([args.dem_launcher, "--srtm", args.srtm, "--swbd", args.swbd, "-p", args.processes_number_dem, "-w", dem_working_dir, args.input, dem_output_dir]) != 0:
+        log(general_log_path, "DEM failed", general_log_filename)
+        sys.exit(-1)
 
 log(general_log_path, "DEM finished in: {}".format(datetime.timedelta(seconds=(time.time() - start))), general_log_filename)
 
@@ -109,10 +117,10 @@ if not create_sym_links(gips, working_dir):
     log(general_log_path, "Symbolic links for GIP files could not be created in the output directory", general_log_filename)
     sys.exit(-1)
 
-log(general_log_path, "dem_output_dir = {}".format(dem_output_dir), general_log_filename)
 dem_hdrs = glob.glob("{}/*.HDR".format(dem_output_dir))
-log(general_log_path, "dem_hdrs = {}".format(dem_hdrs), general_log_filename)
+log(general_log_path, "DEM output directory {} has DEM hdrs = {}".format(dem_output_dir, dem_hdrs), general_log_filename)
 number_of_processed_tiles = 0
+
 for dem_hdr in dem_hdrs:    
     basename, tile_id = get_dem_hdr_info(dem_hdr)
     dem_dir = glob.glob("{0}/{1}.DBL.DIR".format(dem_output_dir, basename))
@@ -123,7 +131,13 @@ for dem_hdr in dem_hdrs:
         log(general_log_path, "Could not create sym links for {0} and {1}".format(dem_hdr, dem_dir[0]), general_log_filename)
         continue
     start = time.time()
-    if run_command(["ssh", "sen2agri-service@{}".format(args.maccs_address), 
+    product_name = ""
+    if args.input.endswith("/"):
+        product_name = os.path.basename(args.input[:len(args.input) - 1])
+    else:
+        product_name = os.path.basename(args.input)
+    sat_id, acquistion_date = get_product_info(product_name)
+    cmd_array = ["ssh", args.maccs_address, 
                     args.maccs_launcher, 
                     "--input", working_dir, 
                     "--TileId", tile_id, 
@@ -131,27 +145,45 @@ for dem_hdr in dem_hdrs:
                     "--mode", "L2INIT", 
                     "--loglevel", "DEBUG",
                     "--enableTest", "false",
-                    "--CheckXMLFilesWithSchema", "false",
-                    "--conf", "UserConfiguration"]) != 0:
+                    "--CheckXMLFilesWithSchema", "false"]
+    if sat_id == 1:
+        #UserConfiguration has to be added for SENTINEL in cmd_array
+        cmd_array += ["--conf", "UserConfiguration"]
+    elif sat_id == 2:
+        #the cmd_array it's enough for LANDSAT8
+        pass
+    else:
+        #unknown product name !
+        log(general_log_path, "Unknown product name: {}".format(product_name), general_log_filename)
+        log(general_log_path, "Continue to the next tile (if present) ", general_log_filename)
+        remove_sym_links([dem_hdr, dem_dir[0]], working_dir)
+        continue
+    log(general_log_path, "sat_id = {} | acq_date = {}".format(sat_id, acquistion_date), general_log_filename)
+    log(general_log_path, "Starting MACCS for {} | TileID: {}!".format(args.input, tile_id), general_log_filename)
+
+    if run_command(cmd_array) != 0:
         log(general_log_path, "MACCS didn't work for {} | TileID: {}!".format(args.input, tile_id), general_log_filename)
     else:
         number_of_processed_tiles += 1
+        if run_command([os.path.dirname(os.path.abspath(__file__)) + "/mosaic_l2a.py", "-i", args.output, "-w", working_dir]) != 0:
+            log(general_log_path, "Mosaic didn't work")
     log(general_log_path, "MACCS finished in: {}".format(datetime.timedelta(seconds=(time.time() - start))), general_log_filename)
     remove_sym_links([dem_hdr, dem_dir[0]], working_dir)
-'''
-try:
-    shutil.rmtree(dem_working_dir)
-except:
-    log(general_log_path, "Couldn't remove the temp dir {}".format(dem_working_dir), general_log_filename)
-try:
-    shutil.rmtree(dem_output_dir)
-except:
-    log(general_log_path, "Couldn't remove the temp dir {}".format(dem_output_dir), general_log_filename)
-try:
-    shutil.rmtree(working_dir)
-except:
-    log(general_log_path, "Couldn't remove the temp dir {}".format(working_dir), general_log_filename)
-'''
+if args.delete_temp == "True":
+    log(general_log_path, "Remove all the temporary files and directory", general_log_filename)
+    try:
+        shutil.rmtree(dem_working_dir)
+    except:
+        log(general_log_path, "Couldn't remove the temp dir {}".format(dem_working_dir), general_log_filename)
+    try:
+        shutil.rmtree(dem_output_dir)
+    except:
+        log(general_log_path, "Couldn't remove the temp dir {}".format(dem_output_dir), general_log_filename)
+    try:
+        shutil.rmtree(working_dir)
+    except:
+        log(general_log_path, "Couldn't remove the temp dir {}".format(working_dir), general_log_filename)
+
 log(general_log_path, "Ended at {}:".format(time.time()), general_log_filename)
 log(general_log_path, "Total execution {}:".format(datetime.timedelta(seconds=(time.time() - general_start))), general_log_filename)
 if number_of_processed_tiles == 0:
